Fix snapshot CI gap, split bug bounty tests, add Terraform remote state, add Stellar queue-depth HPA metric
- Add a "Check for stale snapshots" step to the integration-tests job in
  ci-cd.yml that fails the build if test_snapshots/ drifts after the test
  suite runs, and document the UPDATE_SNAPSHOTS=true cargo test workflow in
  test_snapshots/README.md.

- Split the 45KB tests/bug_bounty_integration.rs into tests/bug_bounty/
  (helpers.rs, lifecycle_tests.rs, duplicate_detection_tests.rs,
  invitation_tests.rs, transition_tests.rs, db_integration_tests.rs) behind
  a tests/bug_bounty/mod.rs re-export, with tests/bug_bounty_integration.rs
  reduced to a single `mod bug_bounty;` entry point. Each module can now be
  run independently, e.g. `cargo test bug_bounty::lifecycle_tests::`. No
  test behavior changed — assertion count, #[tokio::test] count, and helper
  function count all match the original file exactly.

- Configure an S3 + DynamoDB remote backend for terraform/main.tf and add
  infra/terraform/backend.tf for the infra/terraform/ module, both with
  encrypt = true and partial backend configuration (bucket/dynamodb_table
  supplied at `terraform init` time from the TERRAFORM_STATE_BUCKET and
  TERRAFORM_LOCK_TABLE secrets, documented in
  docs/deployment/PRODUCTION_DEPLOYMENT_RUNBOOK.md along with the one-time
  S3/DynamoDB setup and init/apply commands). The two GitHub Actions
  secrets themselves still need to be added by a repo admin.

- Add a custom aframp_stellar_queue_depth HPA metric for the Stellar
  submission engine: a new externalRules entry in
  k8s/monitoring/prometheus-adapter-config.yaml exposes the existing
  stellar_submission_queue_depth gauge (src/stellar/metrics.rs) as an
  external metric, and infra/k8s/hpa.yaml / k8s/hpa/hpa.yaml both gain an
  External metric targeting it (averageValue: 50, i.e. one replica added
  per 50 queued transactions) so batch-mint bursts scale the deployment
  directly instead of waiting on CPU. Both HPAs already had
  scaleDown.stabilizationWindowSeconds: 300 to prevent thrashing. Staging
  load testing of the new metric is a follow-up, not done here.

Closes #798
Closes #799
Closes #800
Closes #801
